requirements/python2.txt
# Python 2 requirements
paypalrestsdk
python-dotenv
python-dateutil
